Analyze the causal graph to determine constraints and objective

The graph must contain edge attributes named "leftside" and "lrconnect" that takes values 0 and 1. Only one edge may have a value 1 for lrconnect. The shiny app returns a graph in this format.

analyze_graph(graph, constraints, effectt)

Arguments

graph
An aaa-igraph-package object that represents a directed acyclic graph

constraints
A vector of character strings that represent the constraints

effectt
A character string that represents the causal effect of interest

Value

A an object of class "linearcausalproblem", which is a list with the following components. This list can be passed to optimize_effect which interfaces with Balke's code. Print and plot methods are also available.

variables

Character vector of variable names of potential outcomes, these start with 'q' to match Balke's notation

parameters

Character vector of parameter names of observed probabilities, these start with 'p' to match Balke's notation

constraints

Character vector of parsed constraints

objective

Character string defining the objective to be optimized in terms of the variables

p.vals

Matrix of all possible values of the observed data vector, corresponding to the list of parameters.

q.vals

Matrix of all possible values of the response function form of the potential outcomes, corresponding to the list of variables.

parsed.query

A nested list containing information on the parsed causal query.

objective.nonreduced

The objective in terms of the original variables, before algebraic variable reduction. The nonreduced variables can be obtained by concatenating the columns of q.vals.

response.functions

List of response functions.

graph

The graph as passed to the function.

Examples

### confounded exposure and outcome
b <- igraph::graph_from_literal(X -+ Y, Ur -+ X, Ur -+ Y)
V(b)$leftside <- c(0,0,0)
V(b)$latent <- c(0,0,1)
E(b)$rlconnect <- E(b)$edge.monotone <- c(0, 0, 0)
analyze_graph(b, constraints = NULL, effectt = "p{Y(X = 1) = 1} - p{Y(X = 0) = 1}")

#> Ready to compute bounds for the effect p{Y(X = 1) = 1} - p{Y(X = 0) = 1} 
#>  Under the assumption encoded in the graph: + 3/3 edges from 9c0709e (vertex names):
#> [1] X ->Y Ur->X Ur->Y
#> No constraints have been specified 
#>  The bounds will be reported in terms of parameters of the form pab_, which represents the probability P(X = a, Y = b). 
#> Additional information is available in the following list elements: [1] "variables"            "parameters"           "constraints"         
#>  [4] "objective"            "p.vals"               "q.vals"              
#>  [7] "parsed.query"         "unparsed.query"       "user.constraints"    
#> [10] "objective.nonreduced" "response.functions"   "graph"